refactor(isolator): write-components-to-capsules
