--- src/environment/component-capsule.ts
@@ -1,1 +1,0 @@
-export class ComponentCapsule {}
